refactor: resolve InspectCode RedundantUsingDirective, PossibleMultipleEnumeration, RedundantCast and eight further rules (#126)
* refactor: resolve InspectCode RedundantUsingDirective

* refactor: resolve InspectCode PossibleMultipleEnumeration

* refactor: resolve InspectCode RedundantAssignment

* refactor: resolve InspectCode RedundantCast

* refactor: resolve InspectCode RedundantNameQualifier

* refactor: resolve InspectCode BadChildStatementIndent

* refactor: resolve InspectCode ConstantNullCoalescingCondition

* refactor: resolve InspectCode GenericEnumeratorNotDisposed

* refactor: resolve InspectCode RedundantEnumerableCastCall

+++ src/PdfPinata.Test/Fonts/DevanagariShapingTests.cs
@@ -163,7 +163,7 @@ public class DevanagariShapingTests
         using var _ = new Shaping();
 
         DrawnText.Glyphs(DrawnText.Page(Conjunct, Font()))
-            .Should().Equal([(int)Shape(Conjunct).Glyphs.Single().GlyphId],
+            .Should().Equal([Shape(Conjunct).Glyphs.Single().GlyphId],
                 "what the renderer writes is what the shaper chose, not a per-character lookup");
     }
